+++ src/test/groovy/ru/yandex/money/gradle/plugins/library/checkversion/CheckVersionSpec.groovy
@@ -27,8 +27,8 @@ class CheckVersionSpec extends AbstractPluginSpec {
         def result = runTasksSuccessfully("dependencies")
 
         then:
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
     }
 
     def "Not found conflict for outer libraries"() {
@@ -53,6 +53,32 @@ class CheckVersionSpec extends AbstractPluginSpec {
         result.standardError.empty
     }
 
+    def "Not found conflict for excluded libraries"() {
+
+        given:
+        buildFile << """
+            dependencies {
+                compile 'ru.yandex.money.common:yamoney-json-utils:2.0.2',
+                        'ru.yandex.money.common:yamoney-xml-utils:3.0.1',
+                        'ru.yandex.money.common:yamoney-xml-utils:4.0.1',
+                        'ru.yandex.money.common:yamoney-json-utils:4.0.3'
+                            
+            }
+            
+            majorVersionChecker {
+                    includeGroupIdPrefixes = ['ru.yamoney', 'ru.yandex.money']
+                    excludeDependencies = ['ru.yandex.money.common:yamoney-json-utils', 'ru.yandex.money.common:yamoney-xml-utils']
+            }
+             
+                """.stripIndent()
+        when:
+        def result = runTasksSuccessfully("dependencies")
+
+        then:
+        !result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
+        !result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
+    }
+
     def "Found conflict for all libraries"() {
 
         given:
@@ -71,8 +97,8 @@ class CheckVersionSpec extends AbstractPluginSpec {
         def result = runTasksSuccessfully("dependencies")
 
         then:
-        result.standardOutput.contains("There is major vesion conflict for dependency=com.google.guava:guava")
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
+        result.standardOutput.contains("There is major version conflict for dependency=com.google.guava:guava")
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
     }
 
 
@@ -100,8 +126,8 @@ class CheckVersionSpec extends AbstractPluginSpec {
         def result = runTasksSuccessfully("dependencies")
 
         then:
-        !(result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-enum-utils"))
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
+        !(result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-enum-utils"))
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-xml-utils")
     }
 
     def "Found conflict, dependencies has major version equals '+'"() {
@@ -119,7 +145,7 @@ class CheckVersionSpec extends AbstractPluginSpec {
         def result = runTasksSuccessfully("dependencies")
 
         then:
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
     }
 
     def "Found conflict, dependencies has version with '+'"() {
@@ -137,6 +163,6 @@ class CheckVersionSpec extends AbstractPluginSpec {
         def result = runTasksSuccessfully("dependencies")
 
         then:
-        result.standardOutput.contains("There is major vesion conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
+        result.standardOutput.contains("There is major version conflict for dependency=ru.yandex.money.common:yamoney-json-utils")
     }
 }